backend/requirements.txt
Django>=4.2
djangorestframework
playwright
beautifulsoup4
lxml
httpx
pandas
openpyxl
